requirements.txt
PyQt6==6.6.1
moderngl==5.10.0
numpy==1.26.3
librosa==0.10.1
Pillow==10.2.0
sounddevice==0.4.6
soundfile==0.12.1
ffmpeg-python==0.2.0
opencv-python==4.9.0.80